refactor: move every component into separate folder
