+++ react-components/src/components/MovieCard/MovieCard.css
@@ -61,7 +61,7 @@
   display: block;
   width: 20px;
   height: 20px;
-  background-image: url("../assets/icons/time.svg");
+  background-image: url("../../assets/icons/time.svg");
   background-repeat: no-repeat;
 }
