+++ react-components/src/components/MovieCard/MovieCard.tsx
