+++ react-components/src/components/MovieCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MovieCard';
