+++ react-components/src/components/NavItem/index.ts
@@ -1,0 +1,1 @@
+export { default } from './NavItem';
