+++ react-components/src/components/SearchBar/SearchBar.css
