+++ react-components/src/components/SearchBar/SearchBar.tsx
