+++ react-components/src/pages/About/About.tsx
@@ -1,4 +1,4 @@
-import Header from 'components/Header';
+import Header from 'components/Header/Header';
 import React from 'react';
 
 export default class About extends React.Component {
