+++ react-components/src/pages/Home/Home.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import './Home.css';
-import MovieCardsList from 'components/MovieCardsList';
-import SearchBar from 'components/SearchBar';
-import Header from 'components/Header';
+import MovieCardsList from 'components/MovieCardsList/MovieCardsList';
+import SearchBar from 'components/SearchBar/SearchBar';
+import Header from 'components/Header/Header';
 
 export default class Home extends React.Component {
   render() {
