+++ react-components/src/pages/PageNotFound/PageNotFound.tsx
@@ -1,4 +1,4 @@
-import Header from 'components/Header';
+import Header from 'components/Header/Header';
 import React from 'react';
 import './PageNotFound.css';
