+++ react-components/src/pages/PageNotFound/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PageNotFound';
